ai-image-search/requirements.txt
fastapi
uvicorn
pillow
requests
supabase
torch
torchvision
ftfy
regex
tqdm
python-dotenv
python-multipart

git+https://github.com/openai/CLIP.git